Move header code to own file
